Working on the input component.
Did the project folder structure and some basic components.
I'm working on the input component so it's reusable.

Stopped at:
Working on the switch cases if there is an input that needs to be a
date. It need to be a nested switch statement, because an input can be
of many types.