Async http call ok
With the form data, there is an http request made with 'axios'. I don't
know why there is a first http request that prints the website html. I
set the url state to be the trigger for the useEffect, but it triggers
before 'url' change. I need to take a look on that.

Pending: Generate a grid with the http response from the previous
request.